+++ diffsync/__init__.py
@@ -214,11 +214,11 @@ class DiffSyncModel(BaseModel):
                 output += f": {child_ids}"
             else:
                 for child_id in child_ids:
-                    child = self.diffsync.get(modelname, child_id)
-                    if not child:
-                        output += f"\n{margin}    {child_id} (details unavailable)"
-                    else:
+                    try:
+                        child = self.diffsync.get(modelname, child_id)
                         output += "\n" + child.str(include_children=include_children, indent=indent + 4)
+                    except ObjectNotFound:
+                        output += f"\n{margin}    {child_id} (ERROR: details unavailable)"
         return output
 
     @classmethod
@@ -428,6 +428,13 @@ class DiffSync:
                     f'Incorrect field name - {value.__name__} has type name "{value.get_type()}", not "{name}"'
                 )
 
+        for name in cls.top_level:
+            if not hasattr(cls, name):
+                raise AttributeError(f'top_level references attribute "{name}" but it is not a class attribute!')
+            value = getattr(cls, name)
+            if not isclass(value) or not issubclass(value, DiffSyncModel):
+                raise AttributeError(f'top_level references attribute "{name}" but it is not a DiffSyncModel subclass!')
+
     def __str__(self):
         """String representation of a DiffSync."""
         if self.type != self.name:
@@ -455,6 +462,8 @@ class DiffSync:
         margin = " " * indent
         output = ""
         for modelname in self.top_level:
+            if output:
+                output += "\n"
             output += f"{margin}{modelname}"
             models = self.get_all(modelname)
             if not models:
@@ -539,7 +548,10 @@ class DiffSync:
         # So let's live with the slightly too high number of branches (14/12) for now.
         log = logger or self._log
         object_class = getattr(self, element.type)
-        obj = self.get(object_class, element.keys)
+        try:
+            obj: Optional[DiffSyncModel] = self.get(object_class, element.keys)
+        except ObjectNotFound:
+            obj = None
         # Get the attributes that actually differ between source and dest
         diffs = element.get_attrs_diffs()
         log = log.bind(
@@ -634,12 +646,16 @@ class DiffSync:
 
     def get(
         self, obj: Union[Text, DiffSyncModel, Type[DiffSyncModel]], identifier: Union[Text, Mapping]
-    ) -> Optional[DiffSyncModel]:
+    ) -> DiffSyncModel:
         """Get one object from the data store based on its unique id.
 
         Args:
             obj: DiffSyncModel class or instance, or modelname string, that defines the type of the object to retrieve
             identifier: Unique ID of the object to retrieve, or dict of unique identifier keys/values
+
+        Raises:
+            ValueError: if obj is a str and identifier is a dict (can't convert dict into a uid str without a model class)
+            ObjectNotFound: if the requested object is not present
         """
         if isinstance(obj, str):
             modelname = obj
@@ -656,13 +672,14 @@ class DiffSync:
         elif object_class:
             uid = object_class.create_unique_id(**identifier)
         else:
-            self._log.warning(
-                f"Tried to look up a {modelname} by identifier {identifier}, "
-                "but don't know how to convert that to a uid string",
+            raise ValueError(
+                f"Invalid args: ({obj}, {identifier}): "
+                f"either {obj} should be a class/instance or {identifier} should be a str"
             )
-            return None
 
-        return self._data[modelname].get(uid)
+        if uid not in self._data[modelname]:
+            raise ObjectNotFound(f"{modelname} {uid} not present in {self.name}")
+        return self._data[modelname][uid]
 
     def get_all(self, obj: Union[Text, DiffSyncModel, Type[DiffSyncModel]]):
         """Get all objects of a given type.
@@ -688,17 +705,20 @@ class DiffSync:
         Args:
             uids: List of unique id / key identifying object in the database.
             obj: DiffSyncModel class or instance, or modelname string, that defines the type of the objects to retrieve
+
+        Raises:
+            ObjectNotFound: if any of the requested UIDs are not found in the store
         """
         if isinstance(obj, str):
             modelname = obj
         else:
             modelname = obj.get_type()
 
-        # TODO: should this raise an exception if any or all of the uids are not found?
         results = []
         for uid in uids:
-            if uid in self._data[modelname]:
-                results.append(self._data[modelname][uid])
+            if uid not in self._data[modelname]:
+                raise ObjectNotFound(f"{modelname} {uid} not present in {self.name}")
+            results.append(self._data[modelname][uid])
         return results
 
     def add(self, obj: DiffSyncModel):
@@ -735,7 +755,7 @@ class DiffSync:
         uid = obj.get_unique_id()
 
         if uid not in self._data[modelname]:
-            raise ObjectNotFound(f"Object {uid} not present")
+            raise ObjectNotFound(f"{modelname} {uid} not present in {self.name}")
 
         if obj.diffsync is self:
             obj.diffsync = None
@@ -745,9 +765,12 @@ class DiffSync:
         if remove_children:
             for child_type, child_fieldname in obj.get_children_mapping().items():
                 for child_id in getattr(obj, child_fieldname):
-                    child_obj = self.get(child_type, child_id)
-                    if child_obj:
+                    try:
+                        child_obj = self.get(child_type, child_id)
                         self.remove(child_obj, remove_children=remove_children)
+                    except ObjectNotFound:
+                        # Since this is "cleanup" code, log an error and continue, instead of letting the exception raise
+                        self._log.error(f"Unable to remove child {child_id} of {modelname} {uid} - not found!")
 
 
 # DiffSyncModel references DiffSync and DiffSync references DiffSyncModel. Break the typing loop: